Move the parser into its own package
