+++ src/test/java/io/burt/jmespath/Compliance.java
@@ -12,6 +12,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import io.burt.jmespath.parser.ParseException;
 import io.burt.jmespath.jackson.JacksonAdapter;
 
 public class Compliance {
